+++ test/Cargo.toml
@@ -27,16 +27,13 @@ features = ["test"]
 path = "../git-trailers/t"
 features = ["test"]
 
-# [dev-dependencies.gitd-lib-test]
-# path = "../cli/gitd-lib/t"
-
 [dev-dependencies.link-git-test]
 path = "../link-git/t"
 features = ["test"]
 
 [dev-dependencies.radicle-surf-test]
 path = "../radicle-surf/t"
-features = ["test"]
+features = ["serialize", "test"]
 
 [dev-dependencies.git-storage-test]
 path = "../git-storage/t"
